mongodb/requirements.txt
dnspython==2.4.2
pymongo==4.5.0
python-decouple==3.8
